+++ dot/rpc/modules/state.go
@@ -188,6 +188,7 @@ func (sm *StateModule) GetKeys(r *http.Request, req *StateStorageKeyRequest, res
 func (sm *StateModule) GetMetadata(r *http.Request, req *StateRuntimeMetadataQuery, res *string) error {
 	// TODO implement change storage trie so that block hash parameter works (See issue #834)
 	metadata, err := sm.coreAPI.GetMetadata()
+	metadata = append([]byte{109, 101, 116, 97}, metadata...)
 	*res = common.BytesToHex(metadata)
 	return err
 }
